Refactored MapOptionsItems
